django_otel_demo/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Deprecated==1.3.1
Django==4.2.27
importlib_metadata==8.5.0
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
sqlparse==0.5.5
typing_extensions==4.13.2
wrapt==2.0.1
zipp==3.20.2
